vllm_on_langgraph_rag/requirements.txt
python-dotenv>=1.0.1
pydantic>=2.8.2
openai>=1.40.0
langchain>=0.2.12
langchain-core>=0.2.38
langchain-openai>=0.1.21
langgraph>=0.2.14
